benchmarks/social-network/social-network-image-filter/requirements.txt
tensorflow-cpu==2.7.0
pillow==9.0.0
minio==7.1.2
